Better estimation of anscestor tree sizes
Previously, the dimensions of the Ancestor Tree d3 SVG were estimated as
```
const width = depth * boxWidth + (depth - 1) * gapX + 2 * padding
const height = boxHeight * 2 ** (depth - 1) + (2 ** (depth - 1) - 1) * gapY
```

Which is to say, effectively (depth + gapX) * boxWidth by (boxHeight + gapY) * 2^(depth - 1).
This is a reasonable estimate if the ancestor tree is fully populated, but is quite
innacurate if the data is more sparse.

This new estimate:
a) ensures that the width is no wider than the total available depth of ancestors available
b) ensures that the height is representative of the actual height of the nodes rendered
c) ensures that the tree itself is centered in the SVG, instead of the home person

It would be better to actually inspect what d3 is rendering and get the height/width from there,
but I couldn't see a great way of doing that.